Tests've shown that "wait" was not working with OpenMPI.
Reason: in OpenMPI, MPI_Request is a pointer to some structure.
MPI_Wait expected arg of type (MPI_Request*), that is - pointer to
pointer to structure.

However, castPtr was silenty casting one pointer to another, wreaking
havoc.

Fix: extra level of pointerness added as needed. Incidentally, code
for "wait" is now similar for both implementations (OpenMPI and
MPICH2) and works (according to tests).